Fully implemented the function of supporting multiple clients simultaneously.
1. Auction items and new bids are updated in real-time on all clients
2. Improved email sending content
3. All filtering methods in Filters can be used
4. Fixed the issue where the registration page cannot automatically redirect to the login page

+++ frontend/src/pages/Browse.jsx
@@ -30,7 +30,7 @@ function Browse() {
       maxPrice: searchParams.get("maxPrice") || "",
       authenticatedOnly: searchParams.get("authenticatedOnly") === "true",
       daysRemaining: searchParams.get("daysRemaining") ? 
-        parseInt(searchParams.get("daysRemaining")) : 5
+        parseFloat(searchParams.get("daysRemaining")) : 5
     };
     
     setFilters(newFilters);
